LayoutTests/wml/variable-reference-invalid-character-expected.txt

CONSOLE MESSAGE: line 0: A variable reference uses invalid syntax.
CONSOLE MESSAGE: line 0: A variable reference uses invalid syntax.
CONSOLE MESSAGE: line 0: A variable reference uses invalid syntax.
CONSOLE MESSAGE: line 0: A variable reference uses invalid syntax.
WML layout tests - using XHTML scripting


Test that invalid variable references aren't detected until variable substitution is executed

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS pElement1.textContent is "Result:$#var PASSED"
PASS pElement2.textContent is "Result: $(v#ar:e)PASSED"
PASS pElement3.textContent is "Result:$(var:#) PASSED"
PASS pElement4.textContent is "Result: PASSED$(var:unescape)"
PASS pElement1.textContent is "Result:$#var PASSED"
PASS pElement2.textContent is "Result: $(v#ar:e)PASSED"
PASS pElement3.textContent is "Result:$(var:#) PASSED"
PASS pElement4.textContent is "Result: PASSED$(var:unescape)"
PASS successfullyParsed is true

TEST COMPLETE


--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
Start test
Result:$#var PASSED
Result: $(v#ar:e)PASSED
Result:$(var:#) PASSED
Result: PASSED$(var:unescape)
